+++ utilities/create-config.py
@@ -44,6 +44,7 @@ def get_snv_indels_config(dirname):
         "ref_id_mapping": join("id_mappings.tsv"),
         "rrna_refflat": join("ucsc_rrna.refFlat"),
         "bed_variant_hotspots": join("hotspots_genome.bed"),
+        "bed_variant_call_regions": join("call_regions.bed"),
         "gtf": join("Homo_sapiens.GRCh38.104.chr.gtf"),
         "annotation_refflat": join("ucsc_gencode.refFlat"),
         "vep_include_consequence": [